x86/modules: Adapt module loading for PIE support
Adapt the module loading process to support PIE relocations. Similar to
other architectures, generate a dynamic GOT for the module if there are
any GOT references in the module. It is possible to do relocation
relaxation like the linker, however, it would require instruction
decoding. Therefore, to simplify the process, keep the GOT reference.

[gavin: fix the conflict with
223b5e57d0d5 mm/execmem, arch: convert remaining overrides of
module_alloc to execmem]

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Gavin Guo <gavinguo@igalia.com>