x86/stackprotector/64: Use per-cpu stack canary if supported by compiler
GCC 8.1 added options that allow the stack canary to be accessed as
"%gs:__stack_chk_guard", effectively turning it into an ordinary per-cpu
variable. For CONFIG_SMP=n, it is not necessary to set up GSBASE as it
could be set as 0. So use the per-cpu stack canary if supported by the
compiler and keep the fixed location code for compatibility with older
compilers.

[gavin: - conflict with 577c134d311b x86/stackprotector: Work around
strict Clang TLS symbol requirements
- change __stack_chk_guard to __ref_stack_chk_guard in arch/x86/Makefile
- add #endif to terminate CONFIG_STACKPROTECTOR]

Co-developed-by: Brian Gerst <brgerst@gmail.com>
Signed-off-by: Brian Gerst <brgerst@gmail.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Gavin Guo <gavinguo@igalia.com>