x86/pie: Make the GOT read-only
Even after hiding all symbol references, the compiler would still
generate some GOT indirect references, resulting in a non-empty GOT when
PIE is enabled. The GOT is fixed up when relocations are applied. It
should be marked as read-only after booting, so the GOT should be put
into the read-only data section.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>