KVM: x86: Implement gpc refresh for guest usage
PVM uses pfncache to share the PVCS structure between the guest and
host. The flag of pfncache of PVCS is initialized as
KVM_GUEST_AND_HOST_USE_PFN because the PVCS is used inside the
vcpu_run() callback, even in the switcher, where the vcpu is in guest
mode. However, there is no real usage for GUEST_USE_PFN, so the request
in mmu_notifier only kicks the vcpu out of guest mode and no refresh is
done before the next vcpu_run(). Therefore, a new request type is
introduced to request the refresh, and a new callback is used to service
the request.

[gavin: fix the conflicts with the following commits:
- 0277022a77a5 ("KVM: x86/mmu: Declare flush_remote_tlbs{_range}() hooks iff HYPERV!=n")
- aa8d1f48d353 ("KVM: x86/mmu: Introduce a quirk to control memslot zap behavior")
- 896046474f8d ("KVM: x86: Introduce kvm_x86_call() to simplify static calls of kvm_x86_ops")
- f4854bf741c4 ("KVM: x86: Replace static_call_cond() with static_call()")
- e366f92ea99e ("KVM: SEV: Support SEV-SNP AP Creation NAE event")
- 721f5b0dda78 ("KVM: pfncache: allow a cache to be activated with a fixed (userspace) HVA")
Also, the patch depends on reverting the following two commits:
- a4bff3df5147 ("KVM: pfncache: remove KVM_GUEST_USES_PFN usage")
- eefb85b3f031 ("KVM: Drop unused @may_block param from gfn_to_pfn_cache_invalidate_start()")]

Suggested-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Gavin Guo <gavinguo@igalia.com>