KVM: x86/PVM: Implement vcpu_load()/vcpu_put() related callbacks
When preparing to switch to the guest, some guest states that only
matter to userspace can be loaded ahead before VM enter. In PVM, guest
segment registers and user return MSRs are loaded into hardware at that
time. Since LDT and IO bitmap are not supported in PVM guests, they are
cleared as well. When preparing to switch to the host in vcpu_put(),
host states are restored.

[gavin: in following commit, the sched_in has been removed from
kvm_x86_ops and merged into vcpu_load:
- 8fbb696a8f53 ("KVM: x86: Fold kvm_arch_sched_in() into kvm_arch_vcpu_load()")
The sched_in implementatino needs to be removed as well to comply with
the upstream interface.]

Signed-off-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Gavin Guo <gavinguo@igalia.com>