KVM: x86/PVM: Implement allowed range checking for #PF
In PVM, guest is only allowed to be running in the reserved virtual
address range provided by the hypervisor. So guest needs to get the
allowed range information from the MSR and the hypervisor needs to check
the fault address and prevent install mapping in the #PF handler.

[gavin: conflict fixes: remove the pvm_sched_in and rename to
pvm_get_feature_msr]

Signed-off-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Gavin Guo <gavinguo@igalia.com>